use query from consumer
